✨ feat: ProductInformation 파일 분리 & 컨텐츠 추가 / 삭제 구현
